// src/types/UserType.ts
export interface UserType {
    id?:           number
    name:          string
    email:         string
    password:      string
    role:          number
    points?:       number
    document?:     string
    team?:         string
    info?:         string
    number?:       string
    birthday?:     string
    code?:         string
}